Add new test modules and remove outdated GraphQL tests
This commit introduces several new test modules for the dashboard API, enhancing the test coverage for various components. Additionally, it removes outdated GraphQL tests that are no longer relevant, streamlining the test suite and improving maintainability. The changes ensure that the testing framework is up-to-date and focused on current functionality.
